Protector/requirements.txt
lief==0.12.3
pycryptodome==3.20.0